+++ hypothesis-python/src/hypothesis/core.py
@@ -93,8 +93,8 @@ from hypothesis.reporting import (
     verbose_report,
     with_reporter,
 )
-from hypothesis.searchstrategy.collections import TupleStrategy
-from hypothesis.searchstrategy.strategies import MappedSearchStrategy, SearchStrategy
+from hypothesis.strategies.internal.collections import TupleStrategy
+from hypothesis.strategies.internal.strategies import MappedSearchStrategy, SearchStrategy
 from hypothesis.statistics import note_engine_for_statistics
 from hypothesis.utils.conventions import infer
 from hypothesis.vendor.pretty import CUnicodeIO, RepresentationPrinter
