+++ hypothesis-python/src/hypothesis/strategies/__init__.py
@@ -17,7 +17,7 @@
 
 from __future__ import absolute_import, division, print_function
 
-from hypothesis.searchstrategy import SearchStrategy
+from hypothesis.strategies.internal import SearchStrategy
 from hypothesis.strategies.internal.core import (
     DataObject,
     _strategies,
