+++ hypothesis-python/src/hypothesis/strategies/internal/__init__.py
@@ -15,4 +15,11 @@
 #
 # END HEADER
 
+"""Package defining SearchStrategy, which is the core type that Hypothesis uses
+to explore data."""
+
 from __future__ import absolute_import, division, print_function
+
+from .strategies import SearchStrategy, check_strategy
+
+__all__ = ["SearchStrategy", "check_strategy"]
