+++ hypothesis-python/src/hypothesis/strategies/internal/flatmapped.py
@@ -19,7 +19,7 @@ from __future__ import absolute_import, division, print_function
 
 from hypothesis.internal.reflection import get_pretty_function_description
 from hypothesis.internal.validation import check_type
-from hypothesis.searchstrategy.strategies import SearchStrategy
+from hypothesis.strategies.internal.strategies import SearchStrategy
 
 
 class FlatMapStrategy(SearchStrategy):
