+++ hypothesis-python/tests/cover/test_conjecture_test_data.py
@@ -33,7 +33,7 @@ from hypothesis.internal.conjecture.data import (
     Status,
     StopTest,
 )
-from hypothesis.searchstrategy.strategies import SearchStrategy
+from hypothesis.strategies.internal.strategies import SearchStrategy
 
 
 @given(st.binary())
