+++ hypothesis-python/tests/cover/test_type_lookup.py
@@ -29,8 +29,8 @@ from hypothesis.errors import (
     ResolutionFailed,
 )
 from hypothesis.internal.compat import PY2, integer_types
-from hypothesis.searchstrategy import types
-from hypothesis.searchstrategy.types import _global_type_lookup
+from hypothesis.strategies.internal import types
+from hypothesis.strategies.internal.types import _global_type_lookup
 from hypothesis.strategies.internal.core import _strategies
 from tests.common.utils import checks_deprecated_behaviour
