+++ hypothesis-python/tests/py3/test_lookup.py
@@ -35,7 +35,7 @@ from hypothesis.internal.compat import (
     integer_types,
     typing_root_type,
 )
-from hypothesis.searchstrategy import types
+from hypothesis.strategies.internal import types
 from hypothesis.strategies import from_type
 from tests.common.debug import find_any, minimal
 from tests.common.utils import fails_with
