+++ hypothesis-python/tests/quality/test_poisoned_lists.py
@@ -26,7 +26,7 @@ import hypothesis.strategies as st
 from hypothesis import settings
 from hypothesis.internal.compat import ceil, hrange
 from hypothesis.internal.conjecture.engine import ConjectureData, ConjectureRunner
-from hypothesis.searchstrategy import SearchStrategy
+from hypothesis.strategies.internal import SearchStrategy
 
 POISON = "POISON"
